// helpup-ng-app/src/app/models/Type.ts
export class Type {
  dog: number;
  cat: number;
  bunny: number;
  asd: number;
  down: number;
  other: number;
  pet: number;
  affection: number;

  constructor() {
    this.dog = 0;
    this.cat = 0;
    this.bunny = 0;
    this.asd = 0;
    this.down = 0;
    this.other = 0;
    this.pet = 0;
    this.affection = 0;
  }
}